fix: fix broken types and imports & refactor some components
