+++ src/components/EmptyValues/StudentDashboard.tsx
@@ -1,5 +1,3 @@
-import Button from "@components/Button";
-import Link from "next/link";
 import { EmptyValueProps } from "./validateEmptyComponent";
 
 const StudentDashboard = ({ data }: EmptyValueProps) => {
