+++ src/components/MentoringLinkCard/MentoringLinkCard.types.ts
@@ -1,16 +1,26 @@
+import { OptionStatus } from "schemas/create_event_output";
+import { Dispatch, SetStateAction } from "react";
+import { ChipVariant } from "@components/Chip/Chip.types";
+import { Pretify } from "types/helpers";
+
 export type Props = {
-  avatar?: string | null | undefined;
+  avatarUrl?: string | null | undefined;
   name: string;
   id?: string;
   job: string;
-  date: any;
-  hour: any;
-  status: string;
+  date: string;
+  hour: string;
+  status: OptionStatus;
   meetingLink?: string;
   eventId: string;
-  onCancel: () => void;
+  onCancel: (...args: any[]) => void;
 };
 
-export type StatusToVariantMap = {
-  [key: string]: "primary" | "secondary" | "tertiary" | "chipCards";
-};
+// export type StatusToVariantMapX = {
+//   [key: string]: "primary" | "secondary" | "tertiary" | "chipCards";
+// };
+
+export type StatusToVariantMap = Record<
+  string,
+  Extract<ChipVariant, "primary" | "secondary" | "tertiary" | "chipCards">
+>;
